--- src/components/ContentCenter/ContentCenter.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import "./ContentCenter.scss";
-
-export default function ContentCenter() {
-    return (
-        <div className="content-center">
-            <h2>ContentCenter...</h2>
-        </div>
-    )
-}
