--- src/components/ContentCenter/ContentCenter.scss
@@ -1,5 +1,0 @@
-@import "../../scss/index.scss";
-
-.content-center {
-    background-color: #ffffff; 
-}
