+++ src/components/ContentCenter/StyleContentCenter.jsx
@@ -1,6 +1,5 @@
 import { makeStyles } from '@material-ui/core/styles';
 
-
 const useStyles = makeStyles((theme) => ({
 
     icon: {
